Changed type of _textureBrickSize from short to int
Current graphics cards have a max texture size of 32768.

OD-4440

+++ src/osgGeo/TexturePlane.cpp
@@ -553,7 +553,7 @@ const osg::Quat& TexturePlaneNode::getRotation() const
 { return _rotation; }
 
 
-void TexturePlaneNode::setTextureBrickSize( short sz, bool strict )
+void TexturePlaneNode::setTextureBrickSize( int sz, bool strict )
 {
     if ( sz > 0 )
 	_textureBrickSize = sz;
@@ -562,7 +562,7 @@ void TexturePlaneNode::setTextureBrickSize( short sz, bool strict )
 }
 
     
-short TexturePlaneNode::getTextureBrickSize() const
+int TexturePlaneNode::getTextureBrickSize() const
 { return _textureBrickSize; }
